+++ app/routers/basic/basic.py
@@ -66,7 +66,6 @@ async def get_feature_changes(
         schema=commons["schema"],
         api_version=commons["api_version"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
     if not result:
         return {
@@ -290,7 +289,6 @@ async def get_arc_audit_values(
         schema=commons["schema"],
         api_version=commons["api_version"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
     return handle_procedure_result(result)