+++ app/routers/epa/dscenario.py
@@ -74,7 +74,6 @@ async def create_dscenario(
         body,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
         api_version=commons["api_version"],
     )
 
@@ -130,7 +129,6 @@ async def select_dscenario(
         where_data={"cur_user": commons["user_id"]},
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
     db_version = await get_db_version(log, commons["db_manager"], schema=commons["schema"])
     return {
@@ -170,7 +168,6 @@ async def delete_dscenario(
         where_data=where_data,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     if not status:
@@ -263,7 +260,6 @@ async def insert_dscenario_objects(
             data=obj,
             schema=commons["schema"],
             user=commons["user_id"],
-            db_role=commons["db_role"],
         )
         all_rows.extend(rows)
 
@@ -326,7 +322,6 @@ async def upsert_dscenario_objects(
             where_data=where_data,
             schema=commons["schema"],
             user=commons["user_id"],
-            db_role=commons["db_role"],
         )
         all_rows.extend(rows)
 
@@ -373,7 +368,6 @@ async def get_dscenario_object(
         parameters=(dscenario_id, object_id),
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     if not rows:
@@ -433,7 +427,6 @@ async def update_dscenario_object(
         where_data=where_data,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     if not rows:
@@ -487,7 +480,6 @@ async def upsert_dscenario_object(
         where_data=where_data,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
     message_text = "Object inserted" if op == "inserted" else "Object updated"
 
@@ -535,7 +527,6 @@ async def delete_dscenario_object(
         where_data=where_data,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     if not status: