+++ app/routers/om/mapzones/dqa.py
@@ -30,7 +30,6 @@ async def get_macrodqas(commons: CommonsDep):
         columns=None,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     db_version = await get_db_version(log, commons["db_manager"], schema=commons["schema"])
@@ -59,7 +58,6 @@ async def get_dqas(commons: CommonsDep):
         columns=None,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     db_version = await get_db_version(log, commons["db_manager"], schema=commons["schema"])