+++ app/routers/om/waterbalance.py
@@ -104,7 +104,6 @@ async def get_waterbalance(
         parameters=parameters,
         schema=commons["schema"],
         user=commons["user_id"],
-        db_role=commons["db_role"],
     )
 
     db_version = await get_db_version(log, commons["db_manager"], schema=commons["schema"])